little change

+++ src/main/java/com/fzu/bbs/services/CommentServices.java
@@ -3,6 +3,6 @@ package com.fzu.bbs.services;
 import com.fzu.bbs.po.Comment;
 
 public interface CommentServices {
-    boolean addComment(String content,int father,Long passageId,String username);
+    boolean addComment(String content,Long father,Long passageId,String username);
     boolean deleteComment(Long commentId);
 }
