+++ src/main/java/com/fzu/bbs/services/PassageServices.java
@@ -4,6 +4,6 @@ import com.fzu.bbs.po.Passage;
 
 public interface PassageServices {
     boolean addPassage(Passage passage);
-    boolean deletePassageById(int id);
+    boolean deletePassageById(Long id);
     Passage getLatestPassage();
 }
