+++ src/main/java/com/fzu/bbs/services/ThemeServices.java
@@ -1,5 +1,5 @@
 package com.fzu.bbs.services;
 
 public interface ThemeServices {
-    boolean addThemeToPassage(int themeId,int passageId);
+    boolean addThemeToPassage(Long themeId,Long passageId);
 }
